requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.2
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
